Cleaning up and adding another metric
